+++ ElvUI_LuckyoneUI/Core/Core.lua
@@ -2,9 +2,10 @@ local L1UI, E, L, V, P, G = unpack(select(2, ...))
 local CH = E:GetModule('Chat')
 local LuckyDT
 
-local print = print
-local format = format
+local format, print = format, print
 
+local _G = _G
+local IsAddOnLoaded = IsAddOnLoaded
 local ReloadUI = ReloadUI
 local SetCVar = SetCVar
 
@@ -39,6 +40,14 @@ function L1UI:VersionCheck()
 	end
 end
 
+-- DB conversion
+function L1UI:ConvertDB()
+	if E.private.L1UI.install_version ~= nil then
+		E.db.L1UI.install_version = E.private.L1UI.install_version
+		E.private.L1UI.install_version = nil
+	end
+end
+
 -- Set UI Scale
 function L1UI:SetupScale()
 	E.global.general.UIScale = 0.71111111111111
@@ -48,20 +57,48 @@ end
 -- General CVars
 function L1UI:SetupCVars()
 
-	-- Set ElvUI CVars first
-	E:SetupCVars(noDisplayMsg)
-
+	-- Core CVars
 	SetCVar('advancedCombatLogging', 1)
+	SetCVar('alwaysShowActionBars', 1)
+	SetCVar('cameraDistanceMaxZoomFactor', 2.6)
 	SetCVar('ffxDeath', 0)
 	SetCVar('ffxGlow', 0)
+	SetCVar('ffxNether', 0)
+	SetCVar('fstack_preferParentKeys', 0)
+	SetCVar('lockActionBars', 1)
 	SetCVar('rawMouseEnable', 1)
+	SetCVar('screenshotQuality', 10)
+	SetCVar('showNPETutorials', 0)
+	SetCVar('showTutorials', 0)
+	SetCVar('threatWarning', 3)
+	SetCVar('UberTooltips', 1)
 
-	-- I want those in Classic and TBC
+	-- Classic and TBC CVars
 	if not E.Retail then
 		SetCVar('autoLootDefault', 1)
 		SetCVar('instantQuestText', 1)
 		SetCVar('profanityFilter', 0)
 	end
+
+	-- My CVars
+	if L1UI.Me then
+		SetCVar('doNotFlashLowHealthWarning', 1)
+		SetCVar('floatingCombatTextCombatDamage', 0)
+		SetCVar('floatingCombatTextCombatHealing', 0)
+		SetCVar('maxFPS', 144)
+		SetCVar('maxFPSBk', 60)
+		SetCVar('maxFPSLoading', 30)
+		SetCVar('nameplateShowOnlyNames', 1)
+		SetCVar('RAIDweatherDensity', 0)
+		SetCVar('showToastOffline', 0)
+		SetCVar('showToastOnline', 0)
+		SetCVar('showToastWindow', 0)
+		SetCVar('SpellQueueWindow', 180)
+		SetCVar('weatherDensity', 0)
+	end
+
+	_G.InterfaceOptionsActionBarsPanelPickupActionKeyDropDown:SetValue('SHIFT')
+	_G.InterfaceOptionsActionBarsPanelPickupActionKeyDropDown:RefreshValue()
 
 	L1UI:Print(L["CVars have been set."])
 end
@@ -70,14 +107,18 @@ end
 function L1UI:NameplateCVars()
 
 	SetCVar('nameplateLargerScale', 1)
+	SetCVar('nameplateLargeTopInset', -1)
 	SetCVar('nameplateMinAlpha', 1)
 	SetCVar('nameplateMinScale', 1)
 	SetCVar('nameplateMotion', 1)
 	SetCVar('nameplateOccludedAlphaMult', 1)
-	SetCVar('nameplateOverlapH', 1)
-	SetCVar('nameplateOverlapV', 1.7)
+	SetCVar('nameplateOtherBottomInset', -1)
+	SetCVar('nameplateOtherTopInset', -1)
+	SetCVar('nameplateOverlapH', 1.1)
+	SetCVar('nameplateOverlapV', 1.8)
 	SetCVar('nameplateSelectedScale', 1)
 	SetCVar('nameplateSelfAlpha', 1)
+	SetCVar('nameplateSelfTopInset', -1)
 
 	SetCVar('UnitNameEnemyGuardianName', 1)
 	SetCVar('UnitNameEnemyMinionName', 1)
@@ -109,13 +150,13 @@ function L1UI:SetupPrivate()
 	E.private.skins.parchmentRemoverEnable = true
 
 	if E.Retail then
-		E.private.install_complete = "12.77"
+		E.private.install_complete = "12.79"
 		E.private.general.totemBar = false
 	elseif E.TBC then
-		E.private.install_complete = "2.44"
+		E.private.install_complete = "2.46"
 		E.private.general.totemBar = true
 	elseif E.Classic then
-		E.private.install_complete = "1.69"
+		E.private.install_complete = "1.71"
 		E.private.general.totemBar = true
 	end
 end
@@ -174,6 +215,7 @@ function L1UI:Cleanup_Cache(addon, type)
 		end
 
 	elseif addon == 'details' then
+		if not IsAddOnLoaded('Details') then return end
 
 		_detalhes.encounter_spell_pool = {}
 		_detalhes.npcid_pool = {}
@@ -181,10 +223,26 @@ function L1UI:Cleanup_Cache(addon, type)
 		_detalhes.spell_school_cache = {}
 
 	elseif addon == 'plater' then
+		if not IsAddOnLoaded('Plater') then return end
 
 		PlaterDB.captured_casts = {}
 		PlaterDB.captured_spells = {}
 		PlaterDB.profiles.Luckyone.npc_cache = {}
+
+	elseif addon == 'rc' then
+		if not IsAddOnLoaded('RCLootCouncil') then return end
+
+		RCLootCouncilDB.global.cache = {}
+		RCLootCouncilDB.global.log = {}
+		RCLootCouncilDB.global.verTestCandidates = {}
+
+	elseif addon == 'mrt' then
+		if not IsAddOnLoaded('MRT') then return end
+
+		VMRT.Encounter.list = {}
+		VMRT.Encounter.names = {}
+		VMRT.ExCD2.gnGUIDs = {}
+		VMRT.Inspect.Soulbinds = {}
 
 	end
 end