+++ ElvUI_LuckyoneUI/Core/Update.lua
@@ -3,12 +3,14 @@ local L1UI, E, L, V, P, G = unpack(select(2, ...))
 -- Update Layout Buttons (Ingame Config)
 function L1UI:UpdateLayout(layout)
 
-	E:StaggeredUpdateAll()
+	E.db["general"]["customGlow"]["color"]["a"] = 1
+	E.db["general"]["customGlow"]["color"]["b"] = 1
+	E.db["general"]["customGlow"]["color"]["g"] = 1
+	E.db["general"]["customGlow"]["color"]["r"] = 1
+	E.db["general"]["customGlow"]["style"] = "Autocast Shine"
+	E.db["general"]["customGlow"]["useColor"] = true
 
-	E.db["tooltip"]["colorAlpha"] = 0.80
-	E.db["tooltip"]["headerFont"] = "Expressway"
-	E.db["tooltip"]["headerFontOutline"] = "OUTLINE"
-	E.db["tooltip"]["headerFontSize"] = 11
+	E:StaggeredUpdateAll()
 
 	L1UI:Print(L["Successfully updated your current layout to LuckyoneUI version "]..L1UI.Version)
 end