+++ ElvUI_LuckyoneUI/Modules/Blizzard.lua
@@ -5,6 +5,11 @@ local _G = _G
 -- Disabled Blizzard Frames
 function L1UI:DisabledFrames()
 
+	if E.private.L1UI.disabledFrames.AlertFrame then
+		_G.AlertFrame:UnregisterAllEvents()
+		E:DisableMover('AlertFrameMover')
+	end
+
 	if E.private.L1UI.disabledFrames.BossBanner and E.Retail then
 		_G.BossBanner:UnregisterAllEvents()
 		E:DisableMover('BossBannerMover')
@@ -12,10 +17,5 @@ function L1UI:DisabledFrames()
 
 	if E.private.L1UI.disabledFrames.ZoneTextFrame then
 		_G.ZoneTextFrame:UnregisterAllEvents()
-	end
-
-	if E.private.L1UI.disabledFrames.AlertFrame then
-		_G.AlertFrame:UnregisterAllEvents()
-		E:DisableMover('AlertFrameMover')
 	end
 end